Split the Animation functions up to make them more manageable. Fixed bug in AddFrameAt.
