+++ v3/src/animation/frame/CheckFrame.js
@@ -1,0 +1,6 @@
+var CheckFrame = function (index)
+{
+    return (index < this.frames.length);
+};
+
+module.exports = CheckFrame;
