+++ v3/src/animation/frame/CompleteAnimation.js
@@ -1,0 +1,11 @@
+var CompleteAnimation = function (component)
+{
+    if (this.hideOnComplete)
+    {
+        component.parent.visible = false;
+    }
+
+    component.stop(true);
+};
+
+module.exports = CompleteAnimation;
