+++ v3/src/animation/frame/GetFrameAt.js
@@ -1,0 +1,6 @@
+var GetFrameAt = function (index)
+{
+    return this.frames[index];
+};
+
+module.exports = GetFrameAt;
