+++ v3/src/animation/frame/GetNextTick.js
@@ -1,0 +1,8 @@
+var GetNextTick = function (component)
+{
+    //  When is the next update due?
+    component.accumulator -= component.nextTick;
+    component.nextTick = component.msPerFrame + component.currentFrame.duration;
+};
+
+module.exports = GetNextTick;
