+++ v3/src/animation/frame/RemoveFrameAt.js
@@ -1,0 +1,10 @@
+var RemoveFrameAt = function (index)
+{
+    this.frames.splice(index, 1);
+
+    this.updateFrameSequence();
+
+    return this;
+};
+
+module.exports = RemoveFrameAt;
